requirements.txt
flask==2.3.3
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
PyMySQL==1.1.0
cryptography==41.0.5
gunicorn==21.2.0